# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pelayanan Publik API",
        "description": "API Backend untuk Aplikasi Pelayanan Publik",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/jenis-layanan/{id}": {
            "get": {
                "tags": [
                    "Jenis Layanan"
                ],
                "summary": "Get single jenis layanan with layanans",
                "operationId": "886d66616597a8dc50da9ff63920e36b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jenis layanan with layanans"
                    }
                }
            }
        },
        "/admin/jenis-layanan": {
            "post": {
                "tags": [
                    "Jenis Layanan"
                ],
                "summary": "Create new jenis layanan",
                "operationId": "e5089cbf0787693cbaa14aa1180400ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nama_jenis": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "icon": {
                                        "type": "file"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jenis layanan created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/jenis-layanan/{id}": {
            "put": {
                "tags": [
                    "Jenis Layanan"
                ],
                "summary": "Update jenis layanan",
                "operationId": "e8b8092f65c65de407352cde5c55b6ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jenis layanan updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jenis Layanan"
                ],
                "summary": "Delete jenis layanan",
                "operationId": "97f044204a3035df5d37520496a7afdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jenis layanan deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}